requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai-whisper==20231117
ffmpeg-python==0.2.0
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
python-dotenv==1.0.0
aiofiles==23.2.1 